pyproject.toml
[tool.poetry]
name = "NONVME_gendiff"
version = "0.4.2"
description = "Json and Yaml files difference generator"
authors = ["NONVME <efimmm.ev@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"
pytest = "^6.1.1"
coverage = "^5.3"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
